+++ doc.go
@@ -36,7 +36,7 @@ Source code and other details for the project are available at GitHub:
 
 Current Version
 
-0.0.11
+0.0.12
 
 Installation
 
@@ -180,20 +180,17 @@ Example Code:
 	func main() {
 
 		// First argument is the raw text for outputs
-		// that are not support colors,
-		// second argument is the full colorful text (yes it can be different if you wish to).
-		//
-		// If the second argument is empty then golog will update the colorful text to the
-		// default color (i.e red on ErrorText) based on the first argument.
+		// second argument is the color code
+		// and the last, variadic argument can be any `kataras/pio.RichOption`, e.g. pio.Background, pio.Underline.
 
 		// Default is "[ERRO]"
-		golog.ErrorText("|ERROR|", "")
+		golog.ErrorText("|ERROR|", 31)
 		// Default is "[WARN]"
-		golog.WarnText("|WARN|", "")
+		golog.WarnText("|WARN|", 32)
 		// Default is "[INFO]"
-		golog.InfoText("|INFO|", "")
+		golog.InfoText("|INFO|", 34)
 		// Default is "[DBUG]"
-		golog.DebugText("|DEBUG|", "")
+		golog.DebugText("|DEBUG|", 33)
 
 		// Business as usual...
 		golog.SetLevel("debug")
@@ -234,10 +231,9 @@ Example Code:
 		var SuccessLevel golog.Level = 6
 		// Register our level, just three fields.
 		golog.Levels[SuccessLevel] = &golog.LevelMetadata{
-			Name:    "success",
-			RawText: "[SUCC]",
-			// ColorfulText (Green Color[SUCC])
-			ColorfulText: "\x1b[32m[SUCC]\x1b[0m",
+			Name:      "success",
+			Title:     "[SUCC]",
+			ColorCode: 32, // Green
 		}
 
 		// create a new golog logger
@@ -394,4 +390,4 @@ Examples:
 package golog // import "github.com/kataras/golog"
 
 // Version is the version string representation of the "golog" package.
-const Version = "0.0.11"
+const Version = "0.0.12"